move link-state logic out of mgmt::requests
